+++ api/tests/opentrons/conftest.py
@@ -952,7 +952,11 @@ def maximal_liquid_class_def() -> LiquidClassSchemaV1:
                                 offset=Coordinate(x=10, y=20, z=30),
                             ),
                             flowRateByVolume=[(1.0, 35.0), (10.0, 24.0), (50.0, 35.0)],
-                            correctionByVolume=[(0.0, 0.0)],
+                            correctionByVolume=[
+                                (0.0, 0.0),
+                                (10.0, -1.0),
+                                (50.0, -10.0),
+                            ],
                             preWet=True,
                             mix=MixProperties(
                                 enable=True, params=MixParams(repetitions=1, volume=50)
@@ -1001,7 +1005,11 @@ def maximal_liquid_class_def() -> LiquidClassSchemaV1:
                                 offset=Coordinate(x=33, y=22, z=11),
                             ),
                             flowRateByVolume=[(1.0, 50.0)],
-                            correctionByVolume=[(0.0, 0.0)],
+                            correctionByVolume=[
+                                (0.0, 0.0),
+                                (10.0, -1.0),
+                                (50.0, -10.0),
+                            ],
                             mix=MixProperties(
                                 enable=True, params=MixParams(repetitions=1, volume=50)
                             ),
@@ -1056,7 +1064,11 @@ def maximal_liquid_class_def() -> LiquidClassSchemaV1:
                                 offset=Coordinate(x=1, y=3, z=2),
                             ),
                             flowRateByVolume=[(50.0, 50.0)],
-                            correctionByVolume=[(0.0, 0.0)],
+                            correctionByVolume=[
+                                (0.0, 0.0),
+                                (10.0, -1.0),
+                                (50.0, -10.0),
+                            ],
                             conditioningByVolume=[(1.0, 5.0), (45.0, 5.0), (50.0, 0.0)],
                             disposalByVolume=[(1.0, 5.0), (45.0, 5.0), (50.0, 0.0)],
                             delay=DelayProperties(
@@ -1216,5 +1228,109 @@ def custom_pip_n_tip_transfer_properties_dict() -> Dict[
                     delay=DelayPropertiesDict(enabled=False),
                 ),
             )
+        }
+    }
+
+
+@pytest.fixture
+def custom_pip_n_tip_transfer_properties_dict_v2() -> Dict[str, Dict[str, Any]]:
+    """A minimal dictionary representation of transfer properties for a custom pipette and tiprack."""
+    return {
+        "a_custom_pipette_type": {
+            "a_custom_tiprack_uri": {
+                "aspirate": {
+                    "aspirate_position": {
+                        "offset": {"x": 1, "y": 2, "z": 3},
+                        "position_reference": "well-bottom",
+                    },
+                    "correction_by_volume": [(0.0, 0.0)],
+                    "delay": {"enable": False},
+                    "flow_rate_by_volume": [(10.0, 40.0), (20.0, 30.0)],
+                    "mix": {"enable": False},
+                    "pre_wet": True,
+                    "retract": {
+                        "air_gap_by_volume": [(5.0, 3.0), (10.0, 4.0)],
+                        "delay": {"enable": False},
+                        "end_position": {
+                            "offset": {"x": 1, "y": 2, "z": 3},
+                            "position_reference": "well-bottom",
+                        },
+                        "speed": 40,
+                        "touch_tip": {"enable": False},
+                    },
+                    "submerge": {
+                        "delay": {"enable": False},
+                        "speed": 100,
+                        "start_position": {
+                            "offset": {"x": 1, "y": 2, "z": 3},
+                            "position_reference": "well-bottom",
+                        },
+                    },
+                },
+                "dispense": {
+                    "dispense_position": {
+                        "offset": {"x": 1, "y": 2, "z": 3},
+                        "position_reference": "well-bottom",
+                    },
+                    "correction_by_volume": [(0.0, 0.0)],
+                    "delay": {"enable": False},
+                    "flow_rate_by_volume": [(10.0, 40.0), (20.0, 30.0)],
+                    "mix": {"enable": False},
+                    "push_out_by_volume": [(10.0, 7.0), (20.0, 10.0)],
+                    "retract": {
+                        "air_gap_by_volume": [(5.0, 3.0), (10.0, 4.0)],
+                        "blowout": {"enable": False},
+                        "delay": {"enable": False},
+                        "end_position": {
+                            "offset": {"x": 1, "y": 2, "z": 3},
+                            "position_reference": "well-bottom",
+                        },
+                        "speed": 40,
+                        "touch_tip": {"enable": False},
+                    },
+                    "submerge": {
+                        "delay": {"enable": False},
+                        "speed": 100,
+                        "start_position": {
+                            "offset": {"x": 1, "y": 2, "z": 3},
+                            "position_reference": "well-bottom",
+                        },
+                    },
+                },
+                "multi_dispense": {
+                    "dispense_position": {
+                        "offset": {"x": 0, "y": 0, "z": 1},
+                        "position_reference": "well-bottom",
+                    },
+                    "flow_rate_by_volume": [(0, 318)],
+                    "correction_by_volume": [(0, 0)],
+                    "delay": {"enabled": False},
+                    "submerge": {
+                        "delay": {"enabled": False},
+                        "speed": 100,
+                        "start_position": {
+                            "offset": {"x": 0, "y": 0, "z": 2},
+                            "position_reference": "well-top",
+                        },
+                    },
+                    "retract": {
+                        "air_gap_by_volume": [(0, 0)],
+                        "delay": {"enabled": False},
+                        "end_position": {
+                            "offset": {"x": 0, "y": 0, "z": 2},
+                            "position_reference": "well-top",
+                        },
+                        "speed": 50,
+                        "touch_tip": {"enabled": False},
+                        "blowout": {
+                            "enabled": True,
+                            "location": "trash",
+                            "flow_rate": 478,
+                        },
+                    },
+                    "conditioning_by_volume": [(0, 0)],
+                    "disposal_by_volume": [(0, 5)],
+                },
+            }
         }
     }