+++ step-generation/src/__tests__/distribute.test.ts
@@ -288,7 +288,7 @@ describe('advanced settings: volume, mix, pre-wet tip, tip touch, tip position',
       volume: 120,
       mixBeforeAspirate: { times: 2, volume: 50 },
       disposalVolume: 12,
-      blowoutLocation: SOURCE_LABWARE,
+      blowoutLocation: 'source_well',
     } as DistributeArgs
     const result = distribute(
       distributeArgs,
@@ -724,13 +724,53 @@ describe('advanced settings: volume, mix, pre-wet tip, tip touch, tip position',
       aspirateHelper('A1', aspirateVol),
       dispenseHelper('A2', volume),
       dispenseHelper('A3', volume),
-      blowoutSingleToSourceA1,
+      {
+        commandType: 'moveToAddressableArea',
+        key: expect.anything(),
+        params: {
+          addressableAreaName: 'movableTrashA3',
+          offset: {
+            x: 0,
+            y: 0,
+            z: 0,
+          },
+          pipetteId: 'p300SingleId',
+        },
+      },
+      {
+        commandType: 'blowOutInPlace',
+        key: expect.anything(),
+        params: {
+          flowRate: 2.3,
+          pipetteId: 'p300SingleId',
+        },
+      },
 
       ...mixCommands,
       aspirateHelper('A1', aspirateVol),
       dispenseHelper('A4', volume),
       dispenseHelper('A5', volume),
-      blowoutSingleToSourceA1,
+      {
+        commandType: 'moveToAddressableArea',
+        key: expect.anything(),
+        params: {
+          addressableAreaName: 'movableTrashA3',
+          offset: {
+            x: 0,
+            y: 0,
+            z: 0,
+          },
+          pipetteId: 'p300SingleId',
+        },
+      },
+      {
+        commandType: 'blowOutInPlace',
+        key: expect.anything(),
+        params: {
+          flowRate: 2.3,
+          pipetteId: 'p300SingleId',
+        },
+      },
     ])
   })